Move local qw store to separate module
