+++ src/qw/local_store/__init__.py
@@ -1,0 +1,1 @@
+"""Local storage for qw."""
