+++ src/qw/remote_repo/factory.py
@@ -1,8 +1,9 @@
 """Factory for making git hosting services."""
 
 from qw.base import QwError
+from qw.local_store.main import get_configuration
 from qw.remote_repo._github import GitHubService
-from qw.remote_repo.service import GitService, Service, get_configuration
+from qw.remote_repo.service import GitService, Service
 
 
 def get_service(conf: dict | None = None) -> GitService:
